Improve the portability of the gitfo_cache code (to 64-bit systems)
This supresses some "conversion from 'size_t' to 'unsigned int',
possible loss of data" warning messages from the MS Visual C/C++
compiler with -Wp64.

+++ src/fileops.c
@@ -133,7 +133,7 @@ void gitfo_free_map(git_map *out)
 /* cached diskio */
 struct gitfo_cache {
 	git_file fd;
-	unsigned int cache_size, pos;
+	size_t cache_size, pos;
 	unsigned char *cache;
 };
